Remove WW-specific fillin template
I am not able to see the purpose of this template. Maybe it used to matter, or evolved into something that we don't really need. There is the "regular" fillin template at line 8135. And that one is good for this too. In fact, what led me here is that this WW-specific one is making blanks that are far too long. They look nicer after I delete this.