Add important pause examples to slideshow
This documents (at least a little) heretofore undocumented behavior of exactly how `pause` and `subslide` work.  I recognize that this may not be the best place for it (since the "samples" are supposed to be "minimal") but I feel like it is a good idea to have it documented somehow, and obviously the sample article is not the place for slideshow stuff.