Recompile a LaTeX image if neccessary
If `Rerun to get` is present in a `.log` file left over form a LaTeX
compile, then the file needs to be recompiled to reach its final state.
This can happen, e.g., for tikz images where it has saved position
information to the `.aux` file.

This commit adds a check of the log file to see if an image needs
recompilation and recompiles if necessary.